requirements.txt
keyboard==0.13.5
pillow==11.0.0
pystray==0.19.5
six==1.16.0
